Fix: Login User with password
